Changes per PR comments
